renamed components to be more clear
