+++ client/src/components/GameStatisticsPane.jsx
